[FEMR-113] Added check for null height in PDF to fix bug
If a patient's height is not provided, it will be displayed as N/A in the
generated PDF of an encounter. Previously, a null height was displayed as 0'0", which
could easily cause confusion.

PDFController now includes value checking for both heights in feet and inches.

Contributed by Robert Massicotte during the CEN5035 course at FSU