[FEMR-151] Add age group and age value validation in triage
Implements error checking on client side to detect a mismatch
between the entered age and age group during triage.

Years and months are used to compute age, and this age is then compared with the given thresholds.
There is an small validation that checks when the user does not provide years or months (so it assigns a default value of 0)

The form cannot be submitted until the inconsistency is solved.

Contributed by Jarrod Moore during the CEN5035 course at FSU