[FEMR-273] Delete Research Only Setting
Removed the seeder functionality that adds "Research Only" as an option in System Settings. Additionally, added a function that scans existing systems for the "Research Only" option, and if found, deletes it.

[FEMR-273] Remove Research Only Setting

Removed the portion of the code that scans for settings with the name "Research Only" and adds a description to it. Since the "Research Only" option has been removed, this piece of code is no longer needed.