Fixed customPost to use newly created methods for adding new medications
previously used setQuantityTotal, which once had functionality to create previously unadded medications, but was refactored by extract function to no longer have that capability. As a result, two InventoryService method calls are needed (one to make the medication, one to set the quantity). As a result, this is also a fix of femr-314