Added read and write permissions to profile foto and encounter folders
Done upon creation of said folders. The goal is to have a permission scheme of -rw------ (permissions 600), which is to only let the user (owner as well, hopefully) read and write the file. Functions from the File class are used to maintain compatibility between windows and unix-based/posix systems. These functions do in essence the same thing, setting read and write permissions only for the owner of the file