[FEMR-348][FEMR-349] Users can edit quantities of medications and delete medications from old trips
femr 348,349. Pull trip id from manageView instead of currentUser

femr/app/femr/common/dtos/CurrentUser.java assumes that a user is only on one trip. As a result, if you take the current user trip id, you only get the most recent one, or the one they are actually on. If you pull from the viewmodel, then you get the trip id of the page you are on