[FEMR-350] Fix undo button not working after deleting medication from inventory
Toggle behaviour no longer extends to InventoryController

Add a route so that instead of calling a method to toggle soft deletion state, the controller knows right away if we are readding or soft-deleting a medication, because the appropriate controller function is called. Changes extend to inventory page js. new method specifically for adding medication in controller. I'd like to remove toggle behaviour from the JS that does the ajax posts at some point, but only after i find a clean way to do it. Toggle logic no longer exists in InventoryController

Fix custom meds not being readded from custom medications page

Added check in inventory Service to see if the medication was in the trip already. If it was not, then create it, otherwise re-add it. Behavior previously was to always create it regardless if it was in the inventory