'Highlighted Errors', fix configs, little work on visual aids
Highlighted Errors - when a test fails, you get the name of the sequential test that failed, the error message, and the stacktrace packaged nicely together in one block of text on the terminal, with a red background and bold white text, so that it's not lost in the stacktrace of everything.

Other related sequential tests do not run, and just output that they failed because a previous test failed.