Merge pull request #790 from FEMR/fix-multiple-chief-complaints
fix(triage): fixed multiple chief complaints not working

Since this line was commented out, the JSONifyChiefComplaints
function didn't store the result and the backend didn't see
any chief complaints. This just re-adds that line.

fix(triage): chief complaint scroll affordance

Before this, users couldn't see that there were
multiple chief complaints (when that was enabled)
because it was hidden in the overflow. This just expands the
height a bit so the user can see the first chief complaint
and scroll to see the rest.